+++ package.json
@@ -32,27 +32,17 @@
     "trailingComma": "es5"
   },
   "dependencies": {
-    "@babel/code-frame": "^7.8.3",
     "chalk": "^3.0.0",
-    "immer": "^5.3.6",
-    "io-ts-types": "^0.5.5",
-    "lodash.camelcase": "^4.3.0",
-    "lodash.kebabcase": "^4.1.1",
     "strip-ansi": "^6.0.0"
   },
   "devDependencies": {
-    "@types/babel__code-frame": "^7.0.1",
     "@types/jest": "^25.1.0",
-    "@types/lodash.camelcase": "^4.3.6",
-    "@types/lodash.kebabcase": "^4.1.6",
     "@types/request": "^2.48.4",
     "docs-ts": "^0.3.4",
     "eslint-plugin-import": "^2.20.1",
     "esm": "^3.2.25",
-    "fp-ts": "^2.4.2",
     "husky": "^4.2.1",
     "infer-types": "^0.0.2",
-    "io-ts": "^2.0.6",
     "jest-junit": "^10.0.0",
     "request": "^2.88.0",
     "tempy": "^0.3.0",
@@ -62,8 +52,5 @@
     "typedoc": "^0.16.10",
     "typescript": "^3.7.5"
   },
-  "peerDependencies": {
-    "fp-ts": "^2.4.2",
-    "io-ts": "^2.0.6"
-  }
+  "peerDependencies": {}
 }
