+++ src/runner.ts
@@ -1,7 +1,7 @@
 import { PrintHelp, Versioned } from './helpdoc';
 import { ParseContext, ParsingResult, Register } from './argparser';
-import { tokenize } from '../newparser/tokenizer';
-import { parse } from '../newparser/parser';
+import { tokenize } from './newparser/tokenizer';
+import { parse } from './newparser/parser';
 import { errorBox } from './errorBox';
 
 export type Handling<Values, Result> = { handler: (values: Values) => Result };
