--- test/fp-ts-helpers.ts
@@ -1,9 +1,0 @@
-import { Either, Right, Left } from 'fp-ts/lib/Either';
-
-export function expectToBeRight(v: Either<any, any>): asserts v is Right<any> {
-  expect(v._tag).toBe('Right');
-}
-
-export function expectToBeLeft(v: Either<any, any>): asserts v is Left<any> {
-  expect(v._tag).toBe('Left');
-}
